refactored UserActivity module declaration
